requirements.txt
numpy>=1.14.0
tensorflow-gpu==1.4.0
opencv-python>=3.2.0
tqdm